# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMDR Therapy Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SyncTherapistRequest {
  name: string;
  email: string;
}

export interface UpdateProfileRequest {
  name: string;
}

export type TherapistProfileStatus =
  (typeof TherapistProfileStatus)[keyof typeof TherapistProfileStatus];

export const TherapistProfileStatus = {
  pending: "pending",
  active: "active",
} as const;

export interface TherapistProfile {
  id: number;
  name: string;
  email: string;
  status: TherapistProfileStatus;
  isAdmin: boolean;
}

export type AdminTherapistViewStatus =
  (typeof AdminTherapistViewStatus)[keyof typeof AdminTherapistViewStatus];

export const AdminTherapistViewStatus = {
  pending: "pending",
  active: "active",
} as const;

export interface AdminTherapistView {
  id: number;
  name: string;
  email: string;
  status: AdminTherapistViewStatus;
  isAdmin: boolean;
  createdAt: string;
}

export interface SessionInfo {
  sessionCode: string;
  therapistId: number;
  expiresAt: string;
}

export interface JoinSessionRequest {
  sessionCode: string;
}

export interface SessionState {
  sessionCode: string;
  isPlaying: boolean;
  speedSeconds: number;
  dotColor: string;
  backgroundColor: string;
}

export interface UpdateSessionStateRequest {
  isPlaying?: boolean;
  speedSeconds?: number;
  dotColor?: string;
  backgroundColor?: string;
}

export interface SavedTheme {
  id: number;
  dotColor: string;
  backgroundColor: string;
  createdAt: string;
}

export interface SaveThemeRequest {
  dotColor: string;
  backgroundColor: string;
}

export interface SendPatientInviteRequest {
  patientEmail: string;
}

export type AdminSearchTherapistsParams = {
  q?: string;
};