# lib/api-zod/src/generated/types/syncTherapistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMDR Therapy Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncTherapistRequest {
  name: string;
  email: string;
}